break scripts into smaller pieces
