+++ smart-contracts/package.json
@@ -50,6 +50,7 @@
     "hardlydifficult-ethereum-contracts": "0.11.1",
     "lint-staged": "11.1.2",
     "prettier": "2.4.0",
+    "prompt": "^1.2.0",
     "truffle-assertions": "0.9.2"
   },
   "devDependencies": {
