+++ smart-contracts/scripts/deployments/uniswap-v2.js
@@ -1,33 +1,18 @@
 const { ethers } = require('hardhat')
 const UniswapV2Factory = require('@uniswap/v2-core/build/UniswapV2Factory.json')
 const UniswapV2Router02 = require('@uniswap/v2-periphery/build/UniswapV2Router02.json')
-const WETH = require('hardlydifficult-eth/src/tokens/weth.json')
-/*
-const WETH = {
-  mainnet: '0xC02aaA39b223FE8D0A0e5C4F27eAD9083C756Cc2',
-  ropsten: '0xc778417E063141139Fce010982780140Aa0cD5Ab',
-  rinkeby: '0xc778417E063141139Fce010982780140Aa0cD5Ab',
-  goerli: '0xB4FBF271143F4FBf7B91A5ded31805e42b2208d6',
-  kovan: '0xd0A1E359811322d97991E03f863a0C30C2cF029C',
-}
-*/
 
 const log = (...message) => {
   // eslint-disable-next-line no-console
   console.log('UNISWAP/WETH SETUP >', ...message)
 }
 
-async function main() {
+async function main({ wethAddress }) {
+  log(`Using WETH contract at: ${wethAddress}`)
+
   const [deployer] = await ethers.getSigners()
   const deployerAddress = deployer.address
   log(`Deploying Uniswap contracts using ${deployerAddress}`)
-
-  // Deploy WETH
-  const weth = await ethers.getContractFactory(WETH.abi, WETH.bytecode)
-  const wethInstance = await weth.deploy()
-  await wethInstance.deployed()
-
-  log(`WETH deployed to : ${wethInstance.address}`)
 
   // Deploy Factory
   const factory = await ethers.getContractFactory(
@@ -46,14 +31,14 @@ async function main() {
   )
   const routerInstance = await router.deploy(
     factoryInstance.address,
-    wethInstance.address
+    wethAddress
   )
   await routerInstance.deployed()
 
   log(`Router V02 deployed to :  ${routerInstance.address}`)
 
   return {
-    weth: wethInstance.address,
+    weth: wethAddress,
     factory: factoryInstance.address,
     router: routerInstance.address,
   }
